+++ plots/plot.encode10.time.sh
@@ -1,11 +1,16 @@
 #!/bin/bash
 
+if [ "$#" != "1" ]; then
+	echo "$0 collectdir"
+	exit
+fi
+
 dir=`pwd`
 
 tmpfile=$dir/tmpfile.R
 rm -rf $tmpfile
 
-sumdir=$dir/encode10/collect.B759
+sumdir=$1
 
 echo "library(\"tikzDevice\")" > $tmpfile
 echo "source(\"$dir/R/encode10.time.R\")" >> $tmpfile
